Improve team assignment support in several locations.
This makes use of changes in mantle-usl for teamAssignments, and does
the proper entity-finds against teamMemberPartyId.  With these changes,
a logged-in user will be able to see Tasks, Projects,
Milestones(inherited via Projects) and Requests that have been only
assigned to a team, that the current user is a member of.